+++ mapper/mapper.go
@@ -17,6 +17,8 @@ import (
 const (
 	InstanceLabel                = dtwebhook.LabelInstance
 	UpdatedViaDynakubeAnnotation = "dynatrace.com/updated-via-operator"
+	ConflictErrorMessage         = "namespace matches two or more DynaKubes which is unsupported. " +
+		"refine the labels on your namespace metadata or DynaKube/CodeModules specification"
 )
 
 type ConflictChecker struct {
@@ -28,8 +30,7 @@ func (c *ConflictChecker) check(dk *dynatracev1beta1.DynaKube) error {
 		return nil
 	}
 	if c.alreadyUsed {
-		return errors.New("namespace matches two or more DynaKubes which is unsupported. " +
-			"refine the labels on your namespace metadata or DynaKube/CodeModules specification")
+		return errors.New(ConflictErrorMessage)
 	}
 	c.alreadyUsed = true
 	return nil
