+++ webhook/validation/namespace_selector.go
@@ -7,9 +7,14 @@ import (
 	"github.com/Dynatrace/dynatrace-operator/mapper"
 )
 
-const errorConflictingNamespaceSelector = `The DynaKube's specification tries to inject into namespaces where another Dynakube already injects into, which is not supported.
+const (
+	errorConflictingNamespaceSelector = `The DynaKube's specification tries to inject into namespaces where another Dynakube already injects into, which is not supported.
 Make sure the namespaceSelector doesn't conflict with other Dynakubes namespaceSelector
 `
+	errorConflictingNamespaceSelectorNoSelector = `The DynaKube does not specificy namespaces where it should inject into while another Dynakube already injects into namespaces, which is not supported.
+Make sure you have a namespaceSelector doesn't conflict with other Dynakubes namespaceSelector
+`
+)
 
 func conflictingNamespaceSelector(dv *dynakubeValidator, dynakube *dynatracev1beta1.DynaKube) string {
 	if !dynakube.NeedAppInjection() {
@@ -17,9 +22,15 @@ func conflictingNamespaceSelector(dv *dynakubeValidator, dynakube *dynatracev1be
 	}
 	dkMapper := mapper.NewDynakubeMapper(context.TODO(), dv.clt, dv.apiReader, dynakube.Namespace, dynakube, log)
 	_, err := dkMapper.MatchingNamespaces()
-	if err != nil {
-		log.Info("requested dynakube has conflicting namespaceSelector", "name", dynakube.Name, "namespace", dynakube.Namespace)
-		return errorConflictingNamespaceSelector
+	if err != nil && err.Error() == mapper.ConflictErrorMessage {
+		if dynakube.NamespaceSelector().MatchExpressions == nil && dynakube.NamespaceSelector().MatchLabels == nil {
+			log.Info("requested dynakube has conflicting namespaceSelector", "name", dynakube.Name, "namespace", dynakube.Namespace)
+			return errorConflictingNamespaceSelectorNoSelector
+		} else {
+			log.Info("requested dynakube has conflicting namespaceSelector", "name", dynakube.Name, "namespace", dynakube.Namespace)
+			return errorConflictingNamespaceSelector
+		}
+
 	}
 	return ""
 }
